move registration java files
